Try to order smaller workflows first in some cases.
This is a small tweak - if two workflows are submitted on the same
day (i.e., they're part of the same batch of jobs, such as we find
at the start of a campaign), then have the smallest workflow go
first.

The intent is to help ramp-up of a new campaign - get as many workflows
started as we can - without causing lots of starvation between old
and new.